requirements.txt
psycopg2-binary==2.9.10
pgvector==0.3.6
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
httpx==0.28.1
jinja2==3.1.6
python-multipart==0.0.20
